TinnitusReconstructor

Documentation for TinnitusReconstructor.

TinnitusReconstructor.unfilled_db

TinnitusReconstructor.Bernoulli

TinnitusReconstructor.BinnedStimgen

TinnitusReconstructor.Brimijoin

TinnitusReconstructor.BrimijoinGaussianSmoothed

TinnitusReconstructor.GaussianNoise

TinnitusReconstructor.GaussianNoiseNoBins

TinnitusReconstructor.GaussianPrior

TinnitusReconstructor.PowerDistribution

TinnitusReconstructor.Stimgen

TinnitusReconstructor.UniformNoise

TinnitusReconstructor.UniformNoiseNoBins

TinnitusReconstructor.UniformPrior

TinnitusReconstructor.UniformPriorWeightedSampling

TinnitusReconstructor.binnedrepr2spect

TinnitusReconstructor.build_distribution

TinnitusReconstructor.crop_signal

TinnitusReconstructor.crop_signal!

TinnitusReconstructor.cs

TinnitusReconstructor.empty_spectrum

TinnitusReconstructor.freq_bins

TinnitusReconstructor.fs

TinnitusReconstructor.generate_stimuli_matrix

TinnitusReconstructor.generate_stimuli_matrix

TinnitusReconstructor.generate_stimulus

TinnitusReconstructor.gs

TinnitusReconstructor.nsamples

TinnitusReconstructor.play_scaled_audio

TinnitusReconstructor.soft_threshold

TinnitusReconstructor.spect2binnedrepr

TinnitusReconstructor.subject_selection_process

TinnitusReconstructor.subject_selection_process

TinnitusReconstructor.subject_selection_process

TinnitusReconstructor.subject_selection_process

TinnitusReconstructor.synthesize_audio

TinnitusReconstructor.wav2spect

TinnitusReconstructor.zhangpassivegamma

TinnitusReconstructor.unfilled_db — Constant

dB level of unfilled bins

TinnitusReconstructor.Bernoulli — Method

Bernoulli(; kwargs...) <: BinnedStimgen

Constructor for stimulus generation type in which each tonotopic bin has a probability bin_prob of being at 0 dB, otherwise it is at -100 dB.

Keywords

min_freq::Real = 100: The minimum frequency in range from which to sample.

max_freq::Real = 22e3: The maximum frequency in range from which to sample.

duration::Real = 0.5: The length of time for which stimuli are played in seconds.

Fs::Real = 44.1e3: The frequency of the stimuli in Hz.

bin_prob::Real=0.3: The probability of a bin being filled.

n_bins::Integer = 100: The number of bins into which to partition the frequency range.

TinnitusReconstructor.BinnedStimgen — Type

BinnedStimgen <: Stimgen

Abstract supertype for all binned stimulus generation.

TinnitusReconstructor.Brimijoin — Method

Brimijoin(; kwargs...) <: BinnedStimgen

Constructor for stimulus generation type in which each tonotopic bin is filled with an amplitude value from an equidistant list with equal probability.

Keywords

min_freq::Real = 100: The minimum frequency in range from which to sample.

max_freq::Real = 22e3: The maximum frequency in range from which to sample.

duration::Real = 0.5: The length of time for which stimuli are played in seconds.

Fs::Real = 44.1e3: The frequency of the stimuli in Hz.

amp_min::Real = -20: The lowest dB value a bin can have.

amp_max::Real = 0: The highest dB value a bin can have.

amp_step::Int = 6: The number of evenly spaced steps between amp_min and amp_max.

n_bins::Integer = 100: The number of bins into which to partition the frequency range.

TinnitusReconstructor.BrimijoinGaussianSmoothed — Method

BrimijoinGaussianSmoothed(; kwargs...) <: BinnedStimgen

Constructor for stimulus generation type in which each tonotopic bin is filled by a Gaussian with a maximum amplitude value chosen from an equidistant list with equal probability.

Keywords

min_freq::Real = 100: The minimum frequency in range from which to sample.

max_freq::Real = 22e3: The maximum frequency in range from which to sample.

duration::Real = 0.5: The length of time for which stimuli are played in seconds.

Fs::Real = 44.1e3: The frequency of the stimuli in Hz.

amp_min::Real = -20: The lowest dB value a bin can have.

amp_max::Real = 0: The highest dB value a bin can have.

amp_step::Int = 6: The number of evenly spaced steps between amp_min and amp_max.

n_bins::Integer = 100: The number of bins into which to partition the frequency range.

TinnitusReconstructor.GaussianNoise — Method

GaussianNoise(; kwargs...) <: BinnedStimgen

Constructor for stimulus generation type in which each tonotopic bin is filled with amplitude chosen from a Gaussian distribution.

Keywords

min_freq::Real = 100: The minimum frequency in range from which to sample.

max_freq::Real = 22e3: The maximum frequency in range from which to sample.

duration::Real = 0.5: The length of time for which stimuli are played in seconds.

Fs::Real = 44.1e3: The frequency of the stimuli in Hz.

n_bins::Integer = 100: The number of bins into which to partition the frequency range.

amplitude_mean::Real = -10: The mean of the Gaussian.

amplitude_var::Real = 3: The variance of the Gaussian.

TinnitusReconstructor.GaussianNoiseNoBins — Method

GaussianNoiseNoBins(; kwargs...) <: Stimgen

Constructor for stimulus generation type in which each frequency's amplitude is chosen according to a Gaussian distribution.

Keywords

min_freq::Real = 100: The minimum frequency in range from which to sample.

max_freq::Real = 22e3: The maximum frequency in range from which to sample.

duration::Real = 0.5: The length of time for which stimuli are played in seconds.

Fs::Real = 44.1e3: The frequency of the stimuli in Hz.

amplitude_mean::Real = -10: The mean of the Gaussian.

amplitude_var::Real = 3: The variance of the Gaussian.

TinnitusReconstructor.GaussianPrior — Method

GaussianPrior(; kwargs...) <: BinnedStimgen

Constructor for stimulus generation type in which the number of filled bins is selected from from a Gaussian distribution with known mean and variance parameters.

Keywords

min_freq::Real = 100: The minimum frequency in range from which to sample.

max_freq::Real = 22e3: The maximum frequency in range from which to sample.

duration::Real = 0.5: The length of time for which stimuli are played in seconds.

Fs::Real = 44.1e3: The frequency of the stimuli in Hz.

n_bins_filled_var::Real = 1: The variance of number of bins that may be filled on any stimuli.

n_bins_filled_mean::Integer = 20: The mean number of bins that may be filled on any stimuli.

n_bins::Integer = 100: The number of bins into which to partition the frequency range.

TinnitusReconstructor.PowerDistribution — Method

PowerDistribution(; kwargs...) <: BinnedStimgen

Constructor for stimulus generation type in which the frequencies in each bin are sampled from a power distribution learned from tinnitus examples.

Keywords

min_freq::Real = 100: The minimum frequency in range from which to sample.

max_freq::Real = 22e3: The maximum frequency in range from which to sample.

duration::Real = 0.5: The length of time for which stimuli are played in seconds.

Fs::Real = 44.1e3: The frequency of the stimuli in Hz.

n_bins::Integer = 100: The number of bins into which to partition the frequency range.

distribution_filepath::AbstractString=joinpath(@__DIR__, "distribution.csv"): The filepath to the default power distribution from which stimuli are generated

TinnitusReconstructor.Stimgen — Type

Abstract supertype for all stimulus generation.

TinnitusReconstructor.UniformNoise — Method

UniformNoise(; kwargs...) <: BinnedStimgen

Constructor for stimulus generation type in which each tonotopic bin is filled with amplitude chosen from a Uniform distribution.

Keywords

min_freq::Real = 100: The minimum frequency in range from which to sample.

max_freq::Real = 22e3: The maximum frequency in range from which to sample.

duration::Real = 0.5: The length of time for which stimuli are played in seconds.

Fs::Real = 44.1e3: The frequency of the stimuli in Hz.

n_bins::Integer = 100: The number of bins into which to partition the frequency range.

TinnitusReconstructor.UniformNoiseNoBins — Method

UniformNoiseNoBins(; kwargs...) <: Stimgen

Constructor for stimulus generation type in which each frequency is chosen from a uniform distribution on [-100, 0] dB.

Keywords

min_freq::Real = 100: The minimum frequency in range from which to sample.

max_freq::Real = 22e3: The maximum frequency in range from which to sample.

duration::Real = 0.5: The length of time for which stimuli are played in seconds.

Fs::Real = 44.1e3: The frequency of the stimuli in Hz.

TinnitusReconstructor.UniformPrior — Method

UniformPrior(; kwargs...) <: BinnedStimgen

Constructor for stimulus generation type in which the number of filled bins is selected from the Uniform distribution on the interval [min_bins, max_bins].

Keywords

min_freq::Real = 100: The minimum frequency in range from which to sample.

max_freq::Real = 22e3: The maximum frequency in range from which to sample.

duration::Real = 0.5: The length of time for which stimuli are played in seconds.

Fs::Real = 44.1e3: The frequency of the stimuli in Hz.

min_bins::Integer = 10: The minimum number of bins that may be filled on any stimuli.

max_bins::Integer = 50: The maximum number of bins that may be filled on any stimuli.

n_bins::Integer = 100: The number of bins into which to partition the frequency range.

TinnitusReconstructor.UniformPriorWeightedSampling — Method

UniformPriorWeightedSampling(; kwargs...) <: BinnedStimgen

Constructor for stimulus generation type in which each tonotopic bin is filled from a uniform distribution on [min_bins, max_bins] but which bins are filled is determined by a non-uniform distribution.

Keywords

min_freq::Real = 100: The minimum frequency in range from which to sample.

max_freq::Real = 22e3: The maximum frequency in range from which to sample.

duration::Real = 0.5: The length of time for which stimuli are played in seconds.

Fs::Real = 44.1e3: The frequency of the stimuli in Hz.

alpha_::Real = 1: The tuning parameter that exponentiates the number of unique frequencies in each bin.

min_bins::Integer = 10: The minimum number of bins that may be filled on any stimuli.

max_bins::Integer = 50: The maximum number of bins that may be filled on any stimuli.

n_bins::Integer = 100: The number of bins into which to partition the frequency range.

TinnitusReconstructor.binnedrepr2spect — Method

binnedrepr2spect(s::BinnedStimgen, binned_repr::AbstractArray{T}) where {BS<:BinnedStimgen,T}

Convert the binned representation into a spectral representation.

Returns an n_frequencies x n_trials spectral array, where n_trials = size(binned_repr, 2).

TinnitusReconstructor.build_distribution — Method

build_distribution(s::PowerDistribution; save_path::AbstractString=@__DIR__)

Builds the default power distribution from ATA tinnitus sample files. Saves the distribution as a vector in dB at save_path.

TinnitusReconstructor.crop_signal! — Method

crop_signal!(audio::SampleBuf{T, I}; start=0, stop=1) where {T, I}

Crops an audio buffer to between start and stop in seconds. TODO: use Unitful to add dimensions to these values.

TinnitusReconstructor.crop_signal — Method

crop_signal(audio::SampleBuf{T, I}; start=0, stop=1) where {T, I}

Returns an audio buffer cropped to between start and stop in seconds. TODO: use Unitful to add dimensions to these values.

See also crop_signal!.

TinnitusReconstructor.cs — Method

cs(responses, Φ, Γ::Integer=32)

One-bit compressed sensing reverse correlation with basis.

References

Plan, Y. and Vershynin, R., 2012. Robust 1-bit compressed sensing and sparse logistic regression: A convex programming approach. IEEE Transactions on Information Theory, 59(1), pp.482-494.

Zhang, L., Yi, J. and Jin, R., 2014, June. Efficient algorithms for robust one-bit compressive sensing. In International Conference on Machine Learning (pp. 820-828). PMLR.

TinnitusReconstructor.empty_spectrum — Method

empty_spectrum(s::BS) where {BS<:BinnedStimgen}

Generate an nfft x 1 vector of Ints, where all values are -100.

TinnitusReconstructor.freq_bins — Method

freq_bins(s::BS) where {BS<:BinnedStimgen}

Generates a vector indicating which frequencies belong to the same bin, following a tonotopic map of audible frequency perception.

TinnitusReconstructor.fs — Method

fs(s::SG) where {SG<:Stimgen}

Return the number of samples per second.

TinnitusReconstructor.generate_stimuli_matrix — Method

generate_stimuli_matrix(s::BS, n_trials::I) where {BS<:BinnedStimgen, I<:Integer}

Generate n_trials of stimuli based on specifications in the stimgen type.

Returns stimuli_matrix, Fs, spect_matrix, binned_repr_matrix.

TinnitusReconstructor.generate_stimuli_matrix — Method

generate_stimuli_matrix(s::SG, n_trials::I) where {SG<:Stimgen, I<:Integer}

Generate n_trials of stimuli based on specifications in the stimgen type.

Returns stimuli_matrix, Fs, spect_matrix, binned_repr_matrix. binned_repr_matrix = nothing if s >: BinnedStimgen.

TinnitusReconstructor.generate_stimulus — Function

generate_stimulus(s::Stimgen)

Generate one stimulus sound.

Returns waveform, sample rate, spectral representation, binned representation, and a frequency vector. Methods are specialized for each concrete subtype of Stimgen.

TinnitusReconstructor.gs — Method

gs(responses, Φ)

Linear reverse correlation.

References

Gosselin, F. and Schyns, P.G., 2003. Superstitious perceptions reveal properties of internal representations. Psychological science, 14(5), pp.505-509.

TinnitusReconstructor.nsamples — Method

nsamples(s::SG) where {SG<:Stimgen}

Return the number of samples as an Integer. This means that the product fs(s) * s.duration must be an Integer or an InexactError will be thrown.

Examples

```jldoctest julia> s = UniformPrior(;Fs=44.1e3, duration=0.5); nsamples(s) 22050

julia> s = UniformPrior(;Fs=44.1e3, duration=0.7); nsamples(s) ERROR: InexactError: Int64(30869.999999999996)

TinnitusReconstructor.play_scaled_audio — Method

play_scaled_audio(x, Fs)

Scales audio signal from -1 to 1 then plays it. Adapted from MATLAB's soundsc().

TinnitusReconstructor.soft_threshold — Method

soft_threshold(α, γ)

Soft thresholding operator for use in compressed sensing.

TinnitusReconstructor.spect2binnedrepr — Method

spect2binnedrepr(s::BinnedStimgen, spect::AbstractVecOrMat{T}) where {BS<:BinnedStimgen,T}

Convert a spectral representation into a binned representation.

Returns an n_trials x n_bins array containing the amplitude of the spectrum in each frequency bin, where n_trials = size(binned_repr, 2). @doc

TinnitusReconstructor.subject_selection_process — Method

subject_selection_process(stimuli::AbstractArray{T}, target_signal::AbstractMatrix{T}) where {T<:Real}

TinnitusReconstructor.subject_selection_process — Method

subject_selection_process(stimuli_matrix::AbstractVecOrMat{T}, target_signal::AbstractVector{T}) where {T<:Real}

Perform the synthetic subject decision process, given a matrix of precomputed stimuli stimuli_matrix and a target_signal. The stimuli_matrix is of size m x n where m is the number of trials and n is the number of samples in the signal. The target_signal is a flat vector of size n or an n x 1 matrix. Return the n-dimensional response vector y as well as the stimuli_matrix as well as nothing for the binned representation.

TinnitusReconstructor.subject_selection_process — Method

subject_selection_process(s::SG, target_signal::AbstractMatrix{T}, n_trials::I) where {SG<:Stimgen,T<:Real,I<:Integer}

TinnitusReconstructor.subject_selection_process — Method

subject_selection_process(s::SG, target_signal::AbstractVector{T}, n_trials::I) where {SG<:Stimgen,T<:Real,I<:Integer}

Perform the synthetic subject decision process, generating the stimuli on-the-fly using the stimulus generation method s.

TinnitusReconstructor.synthesize_audio — Method

synthesize_audio(X, nfft)

Synthesize audio from spectrum, X

TinnitusReconstructor.wav2spect — Method

wav2spect(audio_file::String; duration=0.5)

Load an audio file, crop it to duration, then compute and return the short time Fourier transform.

TinnitusReconstructor.zhangpassivegamma — Method

zhangpassivegamma(Φ::AbstractArray{T}, y, Γ::Integer) where {T<:Real}

Passive algorithm for 1-bit compressed sensing with no basis.

References

Zhang, L., Yi, J. and Jin, R., 2014, June. Efficient algorithms for robust one-bit compressive sensing. In International Conference on Machine Learning (pp. 820-828). PMLR.