Stimgen Methods

All stimgen

TinnitusReconstructor.generate_stimulus — Function

generate_stimulus(s::Stimgen)

Generate one stimulus sound.

Returns waveform, sample rate, spectral representation, binned representation, and a frequency vector (the last two empty if s >: BinnedStimgen). Methods are specialized for each concrete subtype of Stimgen.

TinnitusReconstructor.fs — Function

fs(s::SG) where {SG<:Stimgen}

Return the number of samples per second.

TinnitusReconstructor.nsamples — Function

nsamples(s::SG) where {SG<:Stimgen}

Return the number of samples as an Integer. This means that the product fs(s) * s.duration must be an Integer or an InexactError will be thrown.

Examples

```jldoctest julia> s = UniformPrior(; Fs=44.1e3, duration=0.5); nsamples(s) 22050

julia> s = UniformPrior(; Fs=44.1e3, duration=0.7); nsamples(s) ERROR: InexactError: Int64(30869.999999999996)

TinnitusReconstructor.subject_selection_process — Method

subject_selection_process(s::SG, target_signal::AbstractVector{T}, n_trials::I) where {SG<:Stimgen,T<:Real,I<:Integer}

Perform the synthetic subject decision process, generating the stimuli on-the-fly using the stimulus generation method s.

TinnitusReconstructor.subject_selection_process — Method

subject_selection_process(s::SG, target_signal::AbstractMatrix{T}, n_trials::I) where {SG<:Stimgen,T<:Real,I<:Integer}

Convert target_signal to Vector if passed as a Matrix.

TinnitusReconstructor.generate_stimuli_matrix — Method

generate_stimuli_matrix(s::SG, n_trials::I) where {SG<:Stimgen, I<:Integer}

Generate n_trials of stimuli based on specifications in the stimgen type.

Returns stimuli_matrix, Fs, spect_matrix, binned_repr_matrix. binned_repr_matrix = nothing if s >: BinnedStimgen.

Binned only

TinnitusReconstructor.generate_stimuli_matrix — Method

generate_stimuli_matrix(s::BS, n_trials::I) where {BS<:BinnedStimgen, I<:Integer}

Generate n_trials of stimuli based on specifications in the stimgen type.

Returns stimuli_matrix, Fs, spect_matrix, binned_repr_matrix.

TinnitusReconstructor.freq_bins — Function

freq_bins(s::BS) where {BS<:BinnedStimgen}

Generates a vector indicating which frequencies belong to the same bin, following a tonotopic map of audible frequency perception.

TinnitusReconstructor.empty_spectrum — Function

empty_spectrum(s::BS) where {BS<:BinnedStimgen}

Generate an nfft x 1 vector of Ints, where all values are -100.

TinnitusReconstructor.spect2binnedrepr — Function

spect2binnedrepr(s::BinnedStimgen, spect::AbstractVecOrMat{T}) where {BS<:BinnedStimgen,T}

Convert a spectral representation into a binned representation.

Returns an n_trials x n_bins array containing the amplitude of the spectrum in each frequency bin, where n_trials = size(binned_repr, 2).

TinnitusReconstructor.binnedrepr2spect — Function

binnedrepr2spect(s::BinnedStimgen, binned_repr::AbstractArray{T}) where {BS<:BinnedStimgen,T}

Convert the binned representation into a spectral representation.

Returns an n_frequencies x n_trials spectral array, where n_trials = size(binned_repr, 2).

TinnitusReconstructor.build_distribution — Function

build_distribution(s::PowerDistribution; save_path::AbstractString=@__DIR__)

Builds the default power distribution from ATA tinnitus sample files. Saves the distribution as a vector in dB at save_path.