Stimgen Types

Abstract Types

TinnitusReconstructor.Stimgen — Type

Abstract supertype for all stimulus generation.

TinnitusReconstructor.BinnedStimgen — Type

BinnedStimgen <: Stimgen

Abstract supertype for all binned stimulus generation.

Binned stimgens

TinnitusReconstructor.Bernoulli — Type

Bernoulli(; kwargs...) <: BinnedStimgen

Stimulus generation type in which each tonotopic bin has a probability bin_prob of being at 0 dB, otherwise it is at -100 dB.

Keywords

min_freq::Real = 100: The minimum frequency in range from which to sample.

max_freq::Real = 22e3: The maximum frequency in range from which to sample.

duration::Real = 0.5: The length of time for which stimuli are played in seconds.

Fs::Real = 44.1e3: The frequency of the stimuli in Hz.

bin_prob::Real=0.3: The probability of a bin being filled.

n_bins::Integer = 100: The number of bins into which to partition the frequency range.

TinnitusReconstructor.Brimijoin — Type

Brimijoin(; kwargs...) <: BinnedStimgen

Stimulus generation type in which each tonotopic bin is filled with an amplitude value from an equidistant list with equal probability.

Keywords

min_freq::Real = 100: The minimum frequency in range from which to sample.

max_freq::Real = 22e3: The maximum frequency in range from which to sample.

duration::Real = 0.5: The length of time for which stimuli are played in seconds.

Fs::Real = 44.1e3: The frequency of the stimuli in Hz.

amp_min::Real = -20: The lowest dB value a bin can have.

amp_max::Real = 0: The highest dB value a bin can have.

amp_step::Int = 6: The number of evenly spaced steps between amp_min and amp_max.

n_bins::Integer = 100: The number of bins into which to partition the frequency range.

TinnitusReconstructor.BrimijoinGaussianSmoothed — Type

BrimijoinGaussianSmoothed(; kwargs...) <: BinnedStimgen

Stimulus generation type in which each tonotopic bin is filled by a Gaussian with a maximum amplitude value chosen from an equidistant list with equal probability.

Keywords

min_freq::Real = 100: The minimum frequency in range from which to sample.

max_freq::Real = 22e3: The maximum frequency in range from which to sample.

duration::Real = 0.5: The length of time for which stimuli are played in seconds.

Fs::Real = 44.1e3: The frequency of the stimuli in Hz.

amp_min::Real = -20: The lowest dB value a bin can have.

amp_max::Real = 0: The highest dB value a bin can have.

amp_step::Int = 6: The number of evenly spaced steps between amp_min and amp_max.

n_bins::Integer = 100: The number of bins into which to partition the frequency range.

TinnitusReconstructor.GaussianNoise — Type

GaussianNoise(; kwargs...) <: BinnedStimgen

Stimulus generation type in which each tonotopic bin is filled with amplitude chosen from a Gaussian distribution.

Keywords

min_freq::Real = 100: The minimum frequency in range from which to sample.

max_freq::Real = 22e3: The maximum frequency in range from which to sample.

duration::Real = 0.5: The length of time for which stimuli are played in seconds.

Fs::Real = 44.1e3: The frequency of the stimuli in Hz.

n_bins::Integer = 100: The number of bins into which to partition the frequency range.

amplitude_mean::Real = -10: The mean of the Gaussian.

amplitude_var::Real = 3: The variance of the Gaussian.

TinnitusReconstructor.GaussianPrior — Type

GaussianPrior(; kwargs...) <: BinnedStimgen

Stimulus generation type in which the number of filled bins is selected from from a Gaussian distribution with known mean and variance parameters.

Keywords

min_freq::Real = 100: The minimum frequency in range from which to sample.

max_freq::Real = 22e3: The maximum frequency in range from which to sample.

duration::Real = 0.5: The length of time for which stimuli are played in seconds.

Fs::Real = 44.1e3: The frequency of the stimuli in Hz.

n_bins_filled_var::Real = 1: The variance of number of bins that may be filled on any stimuli.

n_bins_filled_mean::Integer = 20: The mean number of bins that may be filled on any stimuli.

n_bins::Integer = 100: The number of bins into which to partition the frequency range.

TinnitusReconstructor.PowerDistribution — Type

PowerDistribution(; kwargs...) <: BinnedStimgen

Stimulus generation type in which the frequencies in each bin are sampled from a power distribution learned from tinnitus examples.

Keywords

min_freq::Real = 100: The minimum frequency in range from which to sample.

max_freq::Real = 22e3: The maximum frequency in range from which to sample.

duration::Real = 0.5: The length of time for which stimuli are played in seconds.

Fs::Real = 44.1e3: The frequency of the stimuli in Hz.

n_bins::Integer = 100: The number of bins into which to partition the frequency range.

distribution_filepath::AbstractString=joinpath(@__DIR__, "distribution.csv"): The filepath to the default power distribution from which stimuli are generated

TinnitusReconstructor.UniformNoise — Type

UniformNoise(; kwargs...) <: BinnedStimgen

Stimulus generation type in which each tonotopic bin is filled with amplitude chosen from a Uniform distribution.

Keywords

min_freq::Real = 100: The minimum frequency in range from which to sample.

max_freq::Real = 22e3: The maximum frequency in range from which to sample.

duration::Real = 0.5: The length of time for which stimuli are played in seconds.

Fs::Real = 44.1e3: The frequency of the stimuli in Hz.

n_bins::Integer = 100: The number of bins into which to partition the frequency range.

TinnitusReconstructor.UniformPrior — Type

UniformPrior(; kwargs...) <: BinnedStimgen

Stimulus generation type in which the number of filled bins is selected from the Uniform distribution on the interval [min_bins, max_bins].

Keywords

min_freq::Real = 100: The minimum frequency in range from which to sample.

max_freq::Real = 22e3: The maximum frequency in range from which to sample.

duration::Real = 0.5: The length of time for which stimuli are played in seconds.

Fs::Real = 44.1e3: The frequency of the stimuli in Hz.

min_bins::Integer = 10: The minimum number of bins that may be filled on any stimuli.

max_bins::Integer = 50: The maximum number of bins that may be filled on any stimuli.

n_bins::Integer = 100: The number of bins into which to partition the frequency range.

TinnitusReconstructor.UniformPriorWeightedSampling — Type

UniformPriorWeightedSampling(; kwargs...) <: BinnedStimgen

Stimulus generation type in which each tonotopic bin is filled from a uniform distribution on [min_bins, max_bins] but which bins are filled is determined by a non-uniform distribution.

Keywords

min_freq::Real = 100: The minimum frequency in range from which to sample.

max_freq::Real = 22e3: The maximum frequency in range from which to sample.

duration::Real = 0.5: The length of time for which stimuli are played in seconds.

Fs::Real = 44.1e3: The frequency of the stimuli in Hz.

alpha_::Real = 1: The tuning parameter that exponentiates the number of unique frequencies in each bin.

min_bins::Integer = 10: The minimum number of bins that may be filled on any stimuli.

max_bins::Integer = 50: The maximum number of bins that may be filled on any stimuli.

n_bins::Integer = 100: The number of bins into which to partition the frequency range.

Unbinned stimgens

TinnitusReconstructor.UniformNoiseNoBins — Type

UniformNoiseNoBins(; kwargs...) <: Stimgen

Stimulus generation type in which each frequency is chosen from a uniform distribution on [-100, 0] dB.

Keywords

min_freq::Real = 100: The minimum frequency in range from which to sample.

max_freq::Real = 22e3: The maximum frequency in range from which to sample.

duration::Real = 0.5: The length of time for which stimuli are played in seconds.

Fs::Real = 44.1e3: The frequency of the stimuli in Hz.

TinnitusReconstructor.GaussianNoiseNoBins — Type

GaussianNoiseNoBins(; kwargs...) <: Stimgen

Stimulus generation type in which each frequency's amplitude is chosen according to a Gaussian distribution.

Keywords

min_freq::Real = 100: The minimum frequency in range from which to sample.

max_freq::Real = 22e3: The maximum frequency in range from which to sample.

duration::Real = 0.5: The length of time for which stimuli are played in seconds.

Fs::Real = 44.1e3: The frequency of the stimuli in Hz.

amplitude_mean::Real = -10: The mean of the Gaussian.

amplitude_var::Real = 3: The variance of the Gaussian.