TinnitusReconstructor

Documentation for TinnitusReconstructor.

TinnitusReconstructor.BinnedStimgen

TinnitusReconstructor.Stimgen

TinnitusReconstructor.TransformedDense

TinnitusReconstructor.UniformPrior

TinnitusReconstructor.binnedrepr2spect

TinnitusReconstructor.crop_signal

TinnitusReconstructor.crop_signal!

TinnitusReconstructor.cs

TinnitusReconstructor.empty_spectrum

TinnitusReconstructor.freq_bins

TinnitusReconstructor.fs

TinnitusReconstructor.gaussian_kernel

TinnitusReconstructor.generate_stimuli_matrix

TinnitusReconstructor.generate_stimuli_matrix

TinnitusReconstructor.generate_stimulus

TinnitusReconstructor.gs

TinnitusReconstructor.mmd

TinnitusReconstructor.mmd

TinnitusReconstructor.nsamples

TinnitusReconstructor.phase_to_mm

TinnitusReconstructor.play_scaled_audio

TinnitusReconstructor.scaled_uniform

TinnitusReconstructor.soft_threshold

TinnitusReconstructor.spect2binnedrepr

TinnitusReconstructor.stk

TinnitusReconstructor.subject_selection_process

TinnitusReconstructor.synthesize_audio

TinnitusReconstructor.wav2spect

TinnitusReconstructor.zhangpassivegamma

TinnitusReconstructor.BinnedStimgen — Type

BinnedStimgen <: Stimgen

Abstract supertype for all binned stimulus generation.

TinnitusReconstructor.Stimgen — Type

Abstract supertype for all stimulus generation.

TinnitusReconstructor.TransformedDense — Type

TransformedDense(in => out, σ=identity, σ2=identity; bias=true, init=glorot_uniform)
TransformedDense(W::AbstractMatrix, [bias, σ, σ2])

Create a traditional fully connected layer, whose forward pass is given by:
y = σ.(σ2(W) * x .+ bias)

The input x should be a vector of length in, or batch of vectors represented as an in × N matrix, or any array with size(x,1) == in. The out y will be a vector of length out, or a batch with size(y) == (out, size(x)[2:end]...) Keyword bias=false will switch off trainable bias for the layer. The initialisation of the weight matrix is W = init(out, in), calling the function given to keyword init, with default glorot_uniform. The weight matrix and/or the bias vector (of length out) may also be provided explicitly.

This layer differs from Dense in that there is an extra nonlinear transformation of the weight matrix before multiplication with the input.

TinnitusReconstructor.UniformPrior — Method

UniformPrior(; kwargs...) <: BinnedStimgen

Constructor for stimulus generation type in which the number of filled bins is selected from the Uniform distribution on the interval [min_bins, max_bins].

Keywords

min_freq::Real = 100: The minimum frequency in range from which to sample.

max_freq::Real = 22e3: The maximum frequency in range from which to sample.

duration::Real = 0.5: The length of time for which stimuli are played in seconds.

Fs::Real = 44.1e3: The frequency of the stimuli in Hz.

n_bins::Integer = 100: The number of bins into which to partition the frequency range.

min_bins::Integer = 10: The minimum number of bins that may be filled on any stimuli.

max_bins::Integer = 50: The maximum number of bins that may be filled on any stimuli.

TinnitusReconstructor.binnedrepr2spect — Method

binnedrepr2spect(s::BinnedStimgen, binned_repr::AbstractArray{T}) where {BS<:BinnedStimgen,T}

Convert the binned representation into a spectral representation.

Returns an n_frequencies x n_trials spectral array, where n_trials = size(binned_repr, 2).

TinnitusReconstructor.crop_signal! — Method

crop_signal!(audio::SampleBuf{T, I}; start=0, stop=1) where {T, I}

Crops an audio buffer to between start and stop in seconds. TODO: use Unitful to add dimensions to these values.

TinnitusReconstructor.crop_signal — Method

crop_signal(audio::SampleBuf{T, I}; start=0, stop=1) where {T, I}

Returns an audio buffer cropped to between start and stop in seconds. TODO: use Unitful to add dimensions to these values.

See also crop_signal!.

TinnitusReconstructor.cs — Method

cs(responses, Φ, Γ::Integer=32)

One-bit compressed sensing reverse correlation with basis.

References

Plan, Y. and Vershynin, R., 2012. Robust 1-bit compressed sensing and sparse logistic regression: A convex programming approach. IEEE Transactions on Information Theory, 59(1), pp.482-494.

Zhang, L., Yi, J. and Jin, R., 2014, June. Efficient algorithms for robust one-bit compressive sensing. In International Conference on Machine Learning (pp. 820-828). PMLR.

TinnitusReconstructor.empty_spectrum — Method

empty_spectrum(s::BS) where {BS<:BinnedStimgen}

Generate an nfft x 1 vector of Ints, where all values are -100.

TinnitusReconstructor.freq_bins — Method

freq_bins(s::BS) where {BS<:BinnedStimgen}

Generates a vector indicating which frequencies belong to the same bin, following a tonotopic map of audible frequency perception.

TinnitusReconstructor.fs — Method

fs(s::SG) where {SG<:Stimgen}

Return the number of samples per second.

TinnitusReconstructor.gaussian_kernel — Method

gaussian_kernel(x, y; σ=1)

Compute the gaussian kernel for x and y. This is the function

$k_\sigma : \mathbb{R}^{2m} \times \mathbb{R}^{2m} \rightarrow \mathbb{R}, (x, y) \mapsto k_\sigma (x, y) = \exp \left ( - \frac{1}{2\sigma^2} ||x-y||^2 \right )$

Examples

julia> TinnitusReconstructor.gaussian_kernel(1, 1)
1.0

TinnitusReconstructor.generate_stimuli_matrix — Method

generate_stimuli_matrix(s::BS, n_trials::I) where {BS<:BinnedStimgen, I<:Integer}

Generate n_trials of stimuli based on specifications in the stimgen type.

Returns stimuli_matrix, Fs, spect_matrix, binned_repr_matrix.

TinnitusReconstructor.generate_stimuli_matrix — Method

generate_stimuli_matrix(s::SG, n_trials::I) where {SG<:Stimgen, I<:Integer}

Generate n_trials of stimuli based on specifications in the stimgen type.

Returns stimuli_matrix, Fs, spect_matrix, binned_repr_matrix. binned_repr_matrix = nothing if s >: BinnedStimgen.

TinnitusReconstructor.generate_stimulus — Method

generate_stimulus(s::UniformPrior)

Generate one stimulus sound.

Returns waveform, sample rate, spectral representation, binned representation, and a frequency vector.

TinnitusReconstructor.gs — Method

gs(responses, Φ)

Linear reverse correlation.

References

Gosselin, F. and Schyns, P.G., 2003. Superstitious perceptions reveal properties of internal representations. Psychological science, 14(5), pp.505-509.

TinnitusReconstructor.mmd — Function

mmd(x, y, σ=1)

Compute the maximum mean discrepancy (MMD) between x and y using a Gaussian kernel.

Examples

TODO

TinnitusReconstructor.mmd — Method

mmd(x, y; σ=1)

Compute the maximum mean discrepancy (MMD) between x and y using a Gaussian kernel with standard deviation parameter σ.

Examples

julia> mmd(1, 1)
0.0

julia> mmd(1, 2; σ=1)
0.7869386805747332

julia> mmd(1, 2; σ=2)
0.2350061948308091

TinnitusReconstructor.nsamples — Method

nsamples(s::SG) where {SG<:Stimgen}

Return the number of samples as an Integer. This means that the product fs(s) * s.duration must be an Integer or an InexactError will be thrown.

Examples

```jldoctest julia> s = UniformPrior(;Fs=44.1e3, duration=0.5); nsamples(s) 22050

julia> s = UniformPrior(;Fs=44.1e3, duration=0.7); nsamples(s) ERROR: InexactError: Int64(30869.999999999996)

TinnitusReconstructor.phase_to_mm — Method

phase_to_mm(Φ)

Convert a matrix of phases Φ to a measurement matrix via $\frac{1}{\sqrt{m}} \exp(i \Phi)$.

TinnitusReconstructor.play_scaled_audio — Method

play_scaled_audio(x, Fs)

Scales audio signal from -1 to 1 then plays it. Adapted from MATLAB's soundsc().

TinnitusReconstructor.scaled_uniform — Method

scaled_uniform([rng = Flux.default_rng_value()], size...; gain = 1) -> Array
scaled_uniform([rng], kw...) -> Function

Return an Array{Float32} of the given size containing random numbers drawn from a uniform distribution on the interval $gain * [0, 1]$.

Examples

julia> round.(extrema(scaled_uniform(100, 10; gain=2π)), digits=3) (0.004, 6.282)

julia> scaled_uniform(5) |> summary "5-element Vector{Float32}"

TinnitusReconstructor.soft_threshold — Method

soft_threshold(α, γ)

Soft thresholding operator for use in compressed sensing.

TinnitusReconstructor.spect2binnedrepr — Method

spect2binnedrepr(s::BinnedStimgen, spect::AbstractArray{T}) where {BS<:BinnedStimgen,T}

Convert a spectral representation into a binned representation.

Returns an n_trials x n_bins array containing the amplitude of the spectrum in each frequency bin, where n_trials = size(binned_repr, 2). @doc

TinnitusReconstructor.stk — Method

stk(z)

Stack real and imaginary parts of a complex vector z in a real vector stk(z):

$\mathrm{stk} : \mathbb{C}^m \rightarrow \mathbb{R}^{2m}, z \mapsto \mathrm{stk}(z) = \left[\mathcal{R}(z)^{\mathrm{T}}, \mathcal{I}(z)^{\mathrm{T}} \right]^{\mathrm{T}}$

TinnitusReconstructor.subject_selection_process — Function

subject_selection_process(s::SG, target_signal::AbstractVector{T}, n_trials::I) where {SG<:Stimgen, T<:Real, I<:Integer}
subject_selection_process(s::SG, target_signal::AbstractMatrix{T}, n_trials::I) where {SG<:Stimgen, T<:Real, I<:Integer}
subject_selection_process(stimuli::AbstractArray{T}, target_signal::AbstractVector{T}) where {T<:Real}
subject_selection_process(stimuli::AbstractArray{T}, target_signal::AbstractMatrix{T}) where {T<:Real}

Idealized model of a subject performing the task.

Specify a Stimgen type from which to generate stimuli or input a stimuli matrix. If target_signal is a matrix, it must be one dimensional because that method simply applies vec(target_signal). Return an n_trials x 1 or size(stimuli, 1) x 1 vector of -1 for "no" and 1 for "yes".

TinnitusReconstructor.synthesize_audio — Method

synthesize_audio(X, nfft)

Synthesize audio from spectrum, X

TinnitusReconstructor.wav2spect — Method

wav2spect(audio_file::String; duration=0.5)

Load an audio file, crop it to duration, then compute and return the short time Fourier transform.

TinnitusReconstructor.zhangpassivegamma — Method

zhangpassivegamma(Φ::AbstractArray{T}, y, Γ::Integer) where {T<:Real}

Passive algorithm for 1-bit compressed sensing with no basis.

References

Zhang, L., Yi, J. and Jin, R., 2014, June. Efficient algorithms for robust one-bit compressive sensing. In International Conference on Machine Learning (pp. 820-828). PMLR.