TinnitusReconstructor

Documentation for TinnitusReconstructor.

TinnitusReconstructor.BinnedStimgen

TinnitusReconstructor.Stimgen

TinnitusReconstructor.UniformPrior

TinnitusReconstructor.crop_signal

TinnitusReconstructor.crop_signal!

TinnitusReconstructor.cs

TinnitusReconstructor.empty_spectrum

TinnitusReconstructor.freq_bins

TinnitusReconstructor.generate_stimuli_matrix

TinnitusReconstructor.gs

TinnitusReconstructor.play_scaled_audio

TinnitusReconstructor.spect2binnedrepr

TinnitusReconstructor.subject_selection_process

TinnitusReconstructor.synthesize_audio

TinnitusReconstructor.wav2spect

TinnitusReconstructor.BinnedStimgen — Type

BinnedStimgen <: Stimgen

Abstract supertype for all binned stimulus generation.

TinnitusReconstructor.Stimgen — Type

Abstract supertype for all stimulus generation.

TinnitusReconstructor.UniformPrior — Method

UniformPrior(; kwargs...) <: BinnedStimgen

Outer constructor for stimulus generation type in which the number of filled bins is selected from the Uniform distribution on the interval [min_bins, max_bins].

Keywords

min_freq::Real = 100: The minimum frequency in range from which to sample.

max_freq::Real = 22e3: The maximum frequency in range from which to sample.

duration::Real = 0.5: The length of time for which stimuli are played in seconds.

n_trials::Integer = 100: The number of trials the subject will complete.

Fs::Real = 44.1e3: The frequency of the stimuli in Hz.

n_bins::Integer = 100: The number of bins into which to partition the frequency range.

min_bins::Integer = 10: The minimum number of bins that may be filled on any stimuli.

max_bins::Integer = 50: The maximum number of bins that may be filled on any stimuli.

TinnitusReconstructor.crop_signal! — Method

crop_signal!(audio::SampleBuf{T, I}; start=0, stop=1) where {T, I}

Crops an audio buffer to between start and stop in seconds. TODO: use Unitful to add dimensions to these values.

TinnitusReconstructor.crop_signal — Method

crop_signal(audio::SampleBuf{T, I}; start=0, stop=1) where {T, I}

Returns an audio buffer cropped to between start and stop in seconds. TODO: use Unitful to add dimensions to these values.

See also crop_signal!.

TinnitusReconstructor.cs — Function

cs(responses, Phi, Gamma::Integer=32)

Compressed sensing reverse correlation.

TinnitusReconstructor.empty_spectrum — Method

empty_spectrum(s::BinnedStimgen)

Generate an nfft x 1 vector of Ints, where all values are -100.

TinnitusReconstructor.freq_bins — Method

freq_bins(s::BinnedStimgen)

Generates a vector indicating which frequencies belong to the same bin, following a tonotopic map of audible frequency perception.

TinnitusReconstructor.generate_stimuli_matrix — Method

generate_stimuli_matrix(s::Stimgen)

Generate a stimuli matrix based on specifications in the stimgen type.

TinnitusReconstructor.gs — Method

gs(responses, Phi)

Linear reverse correlation.

TinnitusReconstructor.play_scaled_audio — Method

play_scaled_audio(x, Fs)

Scales audio signal from -1 to 1 then plays it. Adapted from MATLAB's soundsc.

TinnitusReconstructor.spect2binnedrepr — Method

spect2binnedrepr(s::BinnedStimgen, spect::AbstractArray{T}) where {T}

Get the binned representation of the spectrum.

Returns a vector containing the amplitude of the spectrum in each frequency bin.

TinnitusReconstructor.subject_selection_process — Function

subject_selection_process(s::Stimgen, target_signal::AbstractVector{T}) where {T<:Real}
subject_selection_process(stimuli::AbstractArray{T}, target_signal::AbstractVector{T}) where {T<:Real}

Idealized model of a subject performing the task.

Specify a Stimgen type from which to generate stimuli or input a stimuli matrix. Return an n_samples x 1 vector of -1 for "no" and 1 for "yes".

TinnitusReconstructor.synthesize_audio — Method

synthesize_audio(X, nfft)

Synthesize audio from spectrum, X

TinnitusReconstructor.wav2spect — Method

wav2spect(audio_file::String; duration=0.5)

Load an audio file, crop it to duration, then compute and return the Welch power spectral density estimate.